move diagnostics.rs into parser/
